舵轮第四次测试
1. 程序自动复位的情况越来越多了, 目前触发情况是在转动自转的时候发生的, 应该还是程序处理来不及, 准备使用官方提供的arm_math库提高函数处理的效率, 避免看门狗复位.
2. 修改了转动状态下右后轮的速度方向.
3. 光电门坏两个了, 有可能是撞击太多坏了.
4. 上了两电池, 准备关注是否还会ODrive供电不稳定.
5. 陀螺仪的修正有时候可以有时候不行, 可能是上电时候的角度太大了, 正处于0与2PI的边界处, 待修改.

+++ New_RTOS_DR/Inc/move.h
@@ -9,7 +9,7 @@ typedef struct DirectSystem
 {
     IMU_Info* IMU;
     PID Dir_Pos_Pid;
-    int16_t Direction_Tar;
+    double Direction_Tar;
     uint8_t Tar_Update;
 }DirectSystem;
